feat: Cambio en el parser (char *cmd)
